+++ rs/system_api/src/sandbox_safe_system_state.rs
@@ -453,7 +453,7 @@ impl SystemStateChanges {
 
         // Verify new certified data isn't too long and set it.
         if let Some(certified_data) = self.new_certified_data.as_ref() {
-            if certified_data.len() > CERTIFIED_DATA_MAX_LENGTH as usize {
+            if certified_data.len() > CERTIFIED_DATA_MAX_LENGTH {
                 return Err(Self::error("Certified data is too large"));
             }
             system_state.certified_data = certified_data.clone();
